test consolidation, test failure fixes, PR comments
