+++ ttnn/cpp/ttnn/operations/data_movement/common/common.hpp
@@ -12,7 +12,7 @@ namespace ttnn {
 namespace operations {
 namespace data_movement {
 
-ttnn::Shape squeeze_shape_to_nd(const ttnn::Shape& output_shape, const uint32_t);
+ttnn::Shape squeeze_shape_to_ND(const ttnn::Shape& output_shape, const uint32_t);
 
 ttnn::Shape squeeze_shape_to_4D(const ttnn::Shape& output_shape);
 ttnn::Shape squeeze_shape_to_3D(const ttnn::Shape& output_shape);
@@ -22,7 +22,7 @@ ttnn::Shape unsqueeze_shape_to_nd(const ttnn::Shape& shape, const uint32_t n);
 ttnn::Shape unsqueeze_shape_to_3D(const ttnn::Shape& shape);
 ttnn::Shape unsqueeze_shape_to_4D(const ttnn::Shape& shape);
 
-ttnn::Shape squeeze_or_unsqueeze_shape_to_nd(const ttnn::Shape& shape, const uint32_t n);
+ttnn::Shape squeeze_or_unsqueeze_shape_to_ND(const ttnn::Shape& shape, const uint32_t n);
 
 uint32_t get_estimated_size_of_cbs(
     const Tensor& input_tensor_a,
